+++ R/tab.R
@@ -102,7 +102,14 @@ tab = function(...
 			class(design$variables[,vr])[1] ))
 	design$variables[,vr] %<>% droplevels
 	if (drop_na) {
+	  idx = which(is.na(levels(design$variables[,vr])))
+	  if (length(idx) > 0) {
+	    idx.ok = which( !(design$variables[,vr] %>% as.numeric %in% idx))
+      design = design[idx.ok,]
+      design$variables[,vr] %<>% droplevels
+	  }
 	  design = design[which(!is.na(design$variables[,vr])),]
+
 	  if(inherits(design, "svyrep.design")) {
 	    design$prob = 1 / design$pweights
 	  }
@@ -134,7 +141,7 @@ tab = function(...
           , ": categorical variable with too many levels: "
           , nlv, ", but ", max_levels
           , " allowed. Try increasing the max_levels argument or "
-          , "see ?set_output"
+          , "see ?set_opts"
           )
 	  return(invisible(NULL))
 	}